Refactor: Change folder structure - src/store/type/card.ts -> src/type/card.ts
